split json_schema into files
